refactor(goesdr): add image extraction module
Add GOES image data extraction module.
